Factor out model_form_util.py.
