+++ hpaction/forms.py
@@ -38,11 +38,5 @@ class FeeWaiverMiscForm(forms.ModelForm):
     asked_before = YesNoRadiosField()
 
 
-class FeeWaiverForm(forms.ModelForm):
-    class Meta:
-        model = FeeWaiverDetails
-        exclude = ['user']
-
-
 class GeneratePDFForm(forms.Form):
     pass
