+++ loc/schema.py
@@ -1,66 +1,12 @@
 from graphql import ResolveInfo
 import graphene
 from graphene_django.types import DjangoObjectType
-from django.forms import ModelForm
 
 from project.util.session_mutation import SessionFormMutation
+from project.util.model_form_util import OneToOneUserModelFormMutation
 from project import slack, schema_registry
 from . import forms, models
 from airtable.sync import sync_user as sync_user_with_airtable
-
-
-class OneToOneUserModelFormMutation(SessionFormMutation):
-    '''
-    A base class that can be used to make any
-    ModelForm that represents a one-to-one relationship
-    with the user into a GraphQL mutation.
-    '''
-
-    class Meta:
-        abstract = True
-
-    login_required = True
-
-    @classmethod
-    def get_form_kwargs(cls, root, info: ResolveInfo, **input):
-        '''
-        Either create a new instance of our model, or get the
-        existing one, and pass it on to the ModelForm.
-        '''
-
-        user = info.context.user
-        model = cls._meta.form_class._meta.model
-        try:
-            instance = model.objects.get(user=user)
-        except model.DoesNotExist:
-            instance = model(user=user)
-        return {"data": input, "instance": instance}
-
-    @classmethod
-    def perform_mutate(cls, form: ModelForm, info: ResolveInfo):
-        '''
-        Save the ModelForm, which will have already been populated with
-        an instance of our model.
-        '''
-
-        form.save()
-        return cls.mutation_success()
-
-    @classmethod
-    def resolve(cls, parent, info: ResolveInfo):
-        '''
-        This can be used as a GraphQL resolver to get the
-        related model instance for the current user.
-        '''
-
-        user = info.context.user
-        if not user.is_authenticated:
-            return None
-        model = cls._meta.form_class._meta.model
-        try:
-            return model.objects.get(user=user)
-        except model.DoesNotExist:
-            return None
 
 
 class AccessDates(SessionFormMutation):
